ISSUE #2387 - move some model settings code out of model routes
