+++ backend/models/invitations.js
@@ -20,7 +20,7 @@ const db = require("../handler/db");
 const User = require("./user");
 const Job = require("./job");
 const Project = require("./project");
-const ModelSetting = require("./modelSetting");
+const { changePermissions, findModelSettings } = require("./modelSetting");
 const systemLogger = require("../logger.js").systemLogger;
 const Mailer = require("../mailer/mailer");
 
@@ -203,7 +203,7 @@ invitations.teamspaceInvitationCheck = async (email, teamspace) => {
 
 const applyModelPermissions = (teamspace, invitedUser, modelsPermissions) => async modelSetting=> {
 	const {permission} = modelsPermissions.find(({model}) => model === modelSetting._id);
-	return await ModelSetting.changePermissions(teamspace, modelSetting._id, modelSetting.permissions.concat({user: invitedUser, permission}));
+	return await changePermissions(teamspace, modelSetting._id, modelSetting.permissions.concat({user: invitedUser, permission}));
 };
 
 const applyProjectPermissions = (teamspace, invitedUser) => async ({ project_admin , project, models}) => {
@@ -211,7 +211,7 @@ const applyProjectPermissions = (teamspace, invitedUser) => async ({ project_adm
 		await Project.setUserAsProjectAdminById(teamspace, project, invitedUser);
 	} else {
 		const modelsIds = models.map(({model}) => model);
-		const modelsList = await ModelSetting.find({ account: teamspace }, {"_id" : {"$in" : modelsIds}});
+		const modelsList = await findModelSettings(teamspace, {"_id" : {"$in" : modelsIds}});
 		await Promise.all(modelsList.map(applyModelPermissions(teamspace, invitedUser, models)));
 	}
 };
